+++ descwl/survey.py
@@ -1,6 +1,8 @@
 """Manage the parameters that define a simulated survey's camera design and observing conditions.
 """
 
+import math
+
 import galsim
 
 class Survey(object):
@@ -55,7 +57,8 @@ class Survey(object):
             self.psf_model = atmospheric_psf_model
         # Create an empty image using (0,0) to index the lower-left corner pixel.
         self.image_bounds = galsim.BoundsI(0,self.image_width-1,0,self.image_height-1)
-        self.image = galsim.Image(bounds = self.image_bounds,scale=self.pixel_scale)
+        self.image = galsim.Image(bounds = self.image_bounds,scale=self.pixel_scale,
+            dtype = np.float64)
 
     def description(self):
         """Describe the survey we simulate.
